Log in Page Trial Code 3
StudyPal Chat App for iOS & Android login and authentication page with Firebase connection. Main logic still being developed and will be uploaded to repository when finished. Background and button design can be edited with SwiftUI. Works best if run on Xcode. This is a contentview.swift file type with many Content Panel objects.